refactor: remove createAt, updatedAt and update .schema
