+++ src/renderer/schemas/Address.schema.ts
@@ -1,7 +1,5 @@
 export interface Address {
-  id: string;
-  name: string;
-  walletAddress: string;
-  createdAt: Date;
-  updatedAt: Date;
+  id?: string;
+  name?: string;
+  walletAddress?: string;
 }
