--- src/renderer/schemas/AddressBook.schemas.ts
@@ -1,11 +1,0 @@
-export interface AddressBook {
-  id?: string;
-  projectDeveloperName?: string;
-  walletAddress?: string;
-}
-
-export interface Address {
-  id?: string;
-  name: string;
-  walletAddress: string;
-}
